+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * [BREAKING CHANGE] Removed Mount.php and moved `mount()` and `unmount()` to MphpD class
 * [BREAKING CHANGE] (re)moved `DB::list_all` and `DB::list_files` into a new `DB::ls` method.
+* [BREAKING CHANGE] `Player::play()` doesn't unpause anymore if no `$pos` is given. `$pos` is now required.
 * [FIX] Fixed bug in parses with `$mode = MPD_CMD_READ_LIST`.
 * [FIX] Return type of `Queue::add_id` is now `int|false` instead of `array`
 * [FIX] Return type of `Queue::add_find` is now `bool` instead of `array`